Make a runnnable playbook for shrinking ceph-disk-made OSDs
shrink_osd.yml does not seem to support OSDs created by ceph-disk
starting from 3.2
Here's what can be used instead if you still have old OSDs.
(They are probably FileStore OSDs, so you'd need to purge then one day)

I originally took shrink_osd.yml that came with ceph-ansible around 2.5
When putting this to github, the closest I found to it was the v3.1.13
So there are a few changes in this commit that I did not intend to do,
but they seem to be all cosmetic.